Refactor Some Changes
