--- lib/interface/home/navigationMenu.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class NavigationMenu extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      
-    );
-  }
-}
